Initial commit after cloning the repo
Added default Git configuration files as a test for the first check-in
after cloning. For some reason, it picks up the "linkinks" subfolder
as a modified file, without any of the files underneath the folder. Hopefully
that does not create another subfolder with the same directory structure as
the current one - fingers crossed!