Removing Feedback link for the chapter view
(See ticket #5 for more details)
While it's a useful link to have throughout the website, we could use as much real estate as possible for showing the book content. We should either move this link inside the book layout, like we do for Full Screen, or not show this link in this view.

Fix: Removed it in chapter view until we figure a better way to surface the Feedback link in Chapter view.